\name{GSED-package}
\alias{GSED-package}
\alias{GSED}
\docType{package}
\title{
Group Sequential Enrichment Design
}
\description{
Provides function to apply "Group sequential enrichment design incorporating subgroup selection" (GSED) method proposed by Magnusson and Turnbull (2013) <doi:10.1002/sim.5738>.
}
\details{
\tabular{ll}{
Package: \tab GSED\cr
Type: \tab Package\cr
Version: \tab 1.6\cr
Date: \tab 2018-03-05\cr
License: \tab GPL-3\cr
}
}
\author{
Marie-Karelle Riviere-Jourdan <eldamjh@gmail.com>
}
\references{
Baldur P. Magnusson and Bruce W. Turnbull. Group sequential enrichment design incorporating subgroup selection. Statistics in Medicine, 2013. <doi:10.1002/sim.5738>
}
